WebFlux Exception Handling --------------------------
When switching to Reactive libraries we removed or commented out a lot of code and setup. Because of this the error handling is broken now.
So, currently if we put a non-existing id in a url for viewing an entity like recipe or ingredient, we get an ugly Internal Server 500 error, so HTTP status 500 instead of the 404 we had before!

For example, we commented out the ExceptionHandler in RecipController as it used ModelAndView and this is linked to Spring MVC, it is servlet specific.
So we can return a view name there instead and add the Model in the argument list like we normally do.

But due to our reactive stack, the Mono is passed into the view layer and actually our custom NotFoundException isn't being triggered in the controller at all. So it won't intervene.
What we do see is a 'org.thymeleaf.exceptions.TemplateInputException: An error happended during template parsing', because the model attribute being passed to the template is null.

[ERROR] [THYMELEAF][boundedElastic-1] Exception processing template "recipes/detail": An error happened during template parsing (template: "class path resource [templates/recipes/detail.html]")
org.thymeleaf.exceptions.TemplateInputException: An error happened during template parsing (template: "class path resource [templates/recipes/detail.html]")

So if we add this exception to the @ExceptionHandler attribute value and make it a list, it should now also handle our error case.

* RecipeController

	package guru.springframework.spring5recipeapp.controller;

	...

	import org.thymeleaf.exceptions.TemplateInputException;

	@Slf4j
	@Controller
	@RequestMapping("/recipes")
	public class RecipeController {

		public static final String VIEWS_400 = "400";
		public static final String VIEWS_404 = "404";

		...

		@ExceptionHandler({ObjectNotFoundException.class, TemplateInputException.class})
		@ResponseStatus(HttpStatus.NOT_FOUND)
		public String handleObjectNotFoundException(Exception exception, Model model) {
			log.warn("Handling ObjectNotFoundException");
			log.warn(exception.getMessage());

			model.addAttribute("exception", exception);

			return "404";
		}
	}